requirements.txt
fastapi==0.115.0
pydantic==2.8.2
uvicorn==0.30.6
pyarrow==16.1.0
PyYAML==6.0.2
fsspec==2024.6.1
gcsfs==2024.6.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.18.2
pillow==10.4.0
pymupdf==1.24.9
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
